+++ src/deploy/AWSConfig.js
@@ -24,6 +24,9 @@ export default class AWSConfig {
       createRoutes: false,
       lambdaFormat: DEFAULT_LAMBDA_FORMAT,
       parameterMgr: ['system', 'secret'],
+      createAuthorizer: '',
+      attachAuthorizer: '',
+      identitySources: ['$request.header.Authorization'],
     });
   }
 
@@ -33,6 +36,9 @@ export default class AWSConfig {
       .withAWSRole(argv.awsRole)
       .withAWSApi(argv.awsApi)
       .withAWSLambdaFormat(argv.awsLambdaFormat)
+      .withAWSCreateAuthorizer(argv.awsCreateAuthorizer)
+      .withAWSAttachAuthorizer(argv.awsAttachAuthorizer)
+      .withAWSIdentitySources(argv.awsIdentitySource)
       .withAWSCleanUpBuckets(argv.awsCleanupBuckets)
       .withAWSCleanUpIntegrations(argv.awsCleanupIntegrations)
       .withAWSCreateRoutes(argv.awsCreateRoutes)
@@ -79,9 +85,26 @@ export default class AWSConfig {
     return this;
   }
 
+  withAWSCreateAuthorizer(value) {
+    this.createAuthorizer = value;
+    return this;
+  }
+
+  withAWSAttachAuthorizer(value) {
+    this.attachAuthorizer = value;
+    return this;
+  }
+
+  withAWSIdentitySources(value) {
+    this.identitySources = value;
+    return this;
+  }
+
   static yarg(yargs) {
     return yargs
-      .group(['aws-region', 'aws-api', 'aws-role', 'aws-cleanup-buckets', 'aws-cleanup-integrations', 'aws-create-routes'], 'AWS Deployment Options')
+      .group(['aws-region', 'aws-api', 'aws-role', 'aws-cleanup-buckets', 'aws-cleanup-integrations',
+        'aws-create-routes', 'aws-create-authorizer', 'aws-attach-authorizer', 'aws-lambda-format',
+        'aws-parameter-manager'], 'AWS Deployment Options')
       .option('aws-region', {
         description: 'the AWS region to deploy lambda functions to',
         type: 'string',
@@ -112,6 +135,23 @@ export default class AWSConfig {
         description: 'Format to use to create lambda functions (note that all dots (\'.\') will be replaced with underscores.',
         type: 'string',
         default: DEFAULT_LAMBDA_FORMAT,
+      })
+      .option('aws-create-authorizer', {
+        description: 'Creates API Gateway authorizer using lambda authorization with this function and the specified name. '
+          + 'The string can contain placeholders (note that all dots (\'.\') are replaced with underscores. '
+          // eslint-disable-next-line no-template-curly-in-string
+          + 'Example: "helix-authorizer_${version}".',
+        type: 'string',
+      })
+      .option('aws-identity-source', {
+        description: 'Identity source to used when creating the authorizer',
+        type: 'string',
+        array: true,
+        default: ['$request.header.Authorization'],
+      })
+      .option('aws-attach-authorizer', {
+        description: 'Attach specified authorizer to routes during linking.',
+        type: 'string',
       })
       .option('aws-cleanup-buckets', {
         description: 'Cleans up stray temporary S3 buckets',