+++ src/template/validate-bundle.js
@@ -9,19 +9,54 @@
  * OF ANY KIND, either express or implied. See the License for the specific language
  * governing permissions and limitations under the License.
  */
-async function run(bundlePath, opts) {
+async function run(bundlePath, cfg) {
   const result = {
+    status: 'ok',
   };
   try {
     const bundle = await import(bundlePath);
-    const main = bundle.default?.main;
-    if (!main || typeof main !== 'function') {
-      throw Error('Action has no main() function.');
+    const lambda = bundle.default?.lambda;
+    if (!lambda || typeof lambda !== 'function') {
+      throw Error('Action has no lambda() function.');
     }
-    if (opts.invoke) {
-      result.response = await main({});
+    const test = cfg.testBundle ?? cfg.test;
+    if (test !== undefined) {
+      const event = {
+        headers: cfg.testHeaders || {},
+        rawPath: `/${cfg.packageName}/${cfg.baseName}/${cfg.version}`,
+        rawQueryString: '',
+        pathParameters: {
+          path: '',
+        },
+        requestContext: {
+          domainName: 'localhost',
+          http: {
+            method: 'get',
+          },
+        },
+      };
+
+      const testUrl = String(test).startsWith('/') ? test : cfg.testUrl;
+      if (testUrl) {
+        const url = new URL(testUrl, 'https://localhost/');
+        event.pathParameters.path = url.pathname.substring(1);
+        event.rawPath += url.pathname;
+        event.rawQueryString = url.searchParams.toString();
+      }
+      const fn = typeof lambda.raw === 'function' ? lambda.raw : lambda;
+      result.response = await fn(event, {
+        invokedFunctionArn: `arn:aws:lambda:us-east-1:118435662149:function:${cfg.packageName}--${cfg.baseName}:${cfg.version}`,
+        getRemainingTimeInMillis: () => 60 * 1000,
+      });
+      if (cfg.verbose) {
+        // eslint-disable-next-line no-console
+        console.log(result.response);
+      }
+      if (result.response.statusCode !== 200) {
+        result.status = 'error';
+        result.error = result.response.statusCode;
+      }
     }
-    result.status = 'ok';
   } catch (e) {
     result.status = 'error';
     result.error = `${e.message}\n${e.stack}`;