+++ test/fixtures/simple-authorizer/index.js
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 Adobe. All rights reserved.
+ * Copyright 2021 Adobe. All rights reserved.
  * This file is licensed to you under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License. You may obtain a copy
  * of the License at http://www.apache.org/licenses/LICENSE-2.0
@@ -9,30 +9,23 @@
  * OF ANY KIND, either express or implied. See the License for the specific language
  * governing permissions and limitations under the License.
  */
-export default class AzureConfig {
-  constructor() {
-    Object.assign(this, {
-      appName: '',
-    });
-  }
 
-  configure(argv) {
-    return this
-      .withAzureApp(argv.azureApp);
-  }
+/* eslint-disable no-console */
 
-  withAzureApp(value) {
-    this.appName = value;
-    return this;
-  }
-
-  static yarg(yargs) {
-    return yargs
-      .group(['azure-app'], 'Azure Deployment Options')
-      .option('azure-app', {
-        description: 'the Azure function app to deploy to',
-        type: 'string',
-        default: '',
-      });
-  }
+/**
+ * Lambda event handler
+ */
+async function handler(event) {
+  return {
+    isAuthorized: event.headers['x-test-authorization'] === 'test',
+    context: {
+      key: 'test',
+    },
+  };
 }
+
+module.exports = {
+  lambda: handler,
+  // make helix-deploy happy
+  main: () => {},
+};